Apply further changes in deploy page
